% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beastclus.R
\name{beastclus}
\alias{beastclus}
\title{Extract Clusters from a BEAST Phylogenetic Tree}
\usage{
beastclus(
  beast_tree,
  metadata,
  post_threshold = 0.7,
  date_range = 90,
  samearea = FALSE
)
}
\arguments{
\item{beast_tree}{A BEAST tree object of class \code{phylo4d} or similar.}

\item{metadata}{A metadata dataframe containing columns \code{label}, \code{location}, and \code{date}.}

\item{post_threshold}{Numeric value specifying the posterior probability threshold for including a cluster (default = 0.70).}

\item{date_range}{Numeric value specifying the maximum allowable date range (in days) within clusters (default = 90).}

\item{samearea}{Logical value indicating whether clusters must originate from the same geographic area (default = FALSE).}
}
\value{
A dataframe summarizing the clusters, including labels, posterior probabilities, areas, date ranges, and the number of tips.
}
\description{
Extract Clusters from a BEAST Phylogenetic Tree
}
\examples{
# Example usage:
# Assuming `beast_tree` is a BEAST tree object and `metadata` is a dataframe
data_csv <- system.file("extdata", "metadata_samp.csv", package = "caIRA")
metadata<-read.csv(data_csv)
data_beast <- system.file("extdata", "pox_strict_comb.tree", package = "caIRA")
beast_tree <- treeio::read.beast(data_beast)
# with the required columns:
beastclus(beast_tree, metadata, post_threshold = 0.50, date_range = 90, samearea = TRUE)
}
