% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caIRA.R
\name{genclus}
\alias{genclus}
\title{Collecting the Cluster Based Tree and Metadata}
\usage{
genclus(tree, metat, bootstrap_threshold, date_range, samearea)
}
\arguments{
\item{tree}{A phylogenetic tree object (likely of class \code{phylo} from the \code{ape} package).}

\item{metat}{A metadata dataframe containing at least the columns \code{label}, \code{location}, and \code{date}.}

\item{bootstrap_threshold}{Numeric value specifying the minimum bootstrap support required for a clade to be considered.}

\item{date_range}{Numeric value specifying the range of dates (in days) within which tips must fall to be considered as part of the same group.}

\item{samearea}{Logical flag indicating whether only groups from the same geographic area should be considered.}
}
\value{
A dataframe containing the monophyletic groups that meet the specified criteria.
}
\description{
This function identifies monophyletic groups within a phylogenetic tree based on provided criteria, such as bootstrap threshold, date range, and geographic location.
}
\examples{
# Load necessary packages
library(ape)
library(dplyr)
library(stringr)
library(progress)

# Generate a random tree with 20 tips
tree <- rtree(n = 20)
# Generate random bootstrap values for each node
set.seed(666)
bootstrap_values <- sample(50:100, size = tree$Nnode, replace = TRUE)
tree$node.label <- bootstrap_values
# Generate metadata
set.seed(666)
areas <- sample(c("Area1", "Area2", "Area3"), size = length(tree$tip.label), replace = TRUE)
start_date <- as.Date("2020-01-01")
end_date <- as.Date("2020-01-20")
random_dates <- sample(seq(start_date, end_date, by = "day"), size = length(tree$tip.label), replace = TRUE)
# Create a data frame with the metadata
metadata <- data.frame(label = tree$tip.label, location = areas, date = random_dates)
# Run the genclus function
genclus(tree, metadata, bootstrap_threshold = 90, date_range = 30, samearea = TRUE)
}
