% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huebreaker.R
\name{huebreaker}
\alias{huebreaker}
\title{Collecting the Cluster Based Tree and Metadata}
\usage{
huebreaker(tree, metat, bootstrap_threshold, date_range, samearea)
}
\arguments{
\item{tree}{tree file}

\item{metat}{metadata file}

\item{bootstrap_threshold}{bootstrap threshold}

\item{date_range}{date range of the tip (days)}

\item{samearea}{same area or not}
}
\value{
a data frame containing the monophyletic groups that meet the criteria
}
\description{
Collecting the Cluster Based Tree and Metadata
}
\examples{
huebreaker(tree, metat, bootstrap_threshold = 90, date_range = 30, samearea = TRUE)
}
