% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huebreaker.R
\name{huebreaker}
\alias{huebreaker}
\title{Collecting the Cluster Based Tree and Metadata}
\usage{
huebreaker(tree, metat, bootstrap_threshold, date_range, samearea)
}
\arguments{
\item{tree}{tree file}

\item{metat}{metadata dataframe contains \code{label}, \code{location}, and \code{date}}

\item{bootstrap_threshold}{Numeric, the bootstrap threshold}

\item{date_range}{Numeric, date range of the tip (days)}

\item{samearea}{Logical, whether to consider only groups from the same area}
}
\value{
a data frame containing the monophyletic groups that meet the criteria
}
\description{
Collecting the Cluster Based Tree and Metadata
}
\examples{
# Load necessary packages
library(ape)
library(dplyr)
library(stringr)
 # Generate a random tree with 20 tips
tree <- rtree(n = 20)
# Generate random bootstrap values for each node
set.seed(666)
bootstrap_values <- sample(50:100, size = tree$Nnode, replace = TRUE)
tree$node.label <- bootstrap_values
# Generate metadata
set.seed(666)
areas <- sample(c("Area1", "Area2", "Area3"), size = length(tree$tip.label), replace = TRUE)
start_date <- as.Date("2020-01-01")
end_date <- as.Date("2020-01-20")
random_dates <- sample(seq(start_date, end_date, by = "day"), size = length(tree$tip.label), replace = TRUE)
# Create a data frame with the metadata
metadata <- data.frame(label = tree$tip.label, location = areas, date = random_dates)
# Run the huebreaker function
huebreaker(tree, metadata, bootstrap_threshold = 90, date_range = 30, samearea = TRUE)
}
